changed types definition for stats repository
