+++ applications/launchpad/gui-react/src/store/containers/types.ts
@@ -1,8 +1,5 @@
 import type { UnlistenFn } from '@tauri-apps/api/event'
 
-import { Dictionary } from '../../types/general'
-import { StatsEntry } from '../../db'
-
 export type ContainerId = string
 
 export enum Container {
@@ -100,14 +97,4 @@ export interface StatsEventPayload {
     }
   }
   networks: Record<string, { tx_bytes: number; rx_bytes: number }>
-}
-
-export interface StatsRepository {
-  add: (
-    network: string,
-    service: Container,
-    secondTimestamp: string,
-    stats: SerializableContainerStats,
-  ) => Promise<void>
-  getGroupedByContainer: (network: string) => Promise<Dictionary<StatsEntry[]>>
 }
